Copy binding and exception handling modules from solentware_misc project.
These need to be available to solentware_grid and solentware_misc, and
possibly solentware_base, projects which are otherwise independent.

I had not noticed the memory leak warning in the tkinter.Misc.bind method
docstring for many years, and ignored it for a while after noticing.

The initial implementation in solentware_misc is in the wrong place, and
does not deal with the problem for tkinter.Canvas and tkinter.Text bind
methods.

These modules will be removed from solentware_misc.